requirements.txt
astroid==1.4.9
backports.functools-lru-cache==1.3
backports.shutil-get-terminal-size==1.0.0
configparser==3.5.0
coverage==4.3.4
decorator==4.0.11
Django==1.8.17
django-appconf==1.0.1
django-autocomplete-light==3.2.1
django-bootstrap3==8.1.0
django-compat==1.0.13
django-datatable==0.2.1
django-debug-toolbar==1.6
django-extensions==1.7.6
django-favicon-plus==0.0.5
-e git+https://github.com/coleifer/django-generic-m2m.git@83b90cd38275b04413f944fe37e4446e4fe09bf6#egg=django_generic_m2m
django-gm2m==0.4.8
django-image-cropping==1.0.4
django-model-utils==2.6.1
django-pipeline==1.6.11
django-pipeline-compass==0.1.5
django-querysetsequence==0.7
django-redis==4.7.0
django-redis-cache==1.7.1
django-selectable==0.9.0
django-shell-plus==1.1.7
django-tagging==0.4.5
django-taggit==0.22.0
easy-thumbnails==2.3
enum34==1.1.6
freeze==1.0.10
funcsigs==1.0.2
futures==3.0.5
gunicorn==19.7.1
html5lib==1.0b8
ipython==5.1.0
ipython-genutils==0.1.0
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.6.1
mock==2.0.0
olefile==0.44
packaging==16.8
pathlib==1.0.1
pathlib2==2.2.1
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
py==1.4.32
Pygments==2.1.3
pylint==1.6.5
pyparsing==2.1.10
PyPDF2==1.26.0
pyScss==1.3.5
pytest==3.0.6
pytest-cov==2.4.0
pytest-django==3.1.2
pytest-splinter==1.8.1
redis==2.10.5
reportlab==3.3.0
requests==2.13.0
scandir==1.4
scss==0.8.73
selenium==3.0.2
simplegeneric==0.8.1
six==1.10.0
sorl-thumbnail==12.3
splinter==0.7.5
sqlparse==0.2.2
traitlets==4.3.1
wcwidth==0.1.7
webencodings==0.5
wrapt==1.10.8
xhtml2pdf==0.0.6
dj-database-url==0.4.1
whitenoise==3.2
